\docType{data}
\name{wrs1_asc_desc}
\alias{wrs1_asc_desc}
\title{Worldwide Reference System - 1 (WRS1) SpatialPolygonsDataFrame}
\description{
Contains polygons of the area of each path and row for the
ascending and descending nodes of the WRS1 system.
}
\references{
\href{http://landsat.usgs.gov/tools_wrs-2_shapefile.php}{WRS-1
Shape files courtesy of the U.S. Geological Survey}.
}
\seealso{
wrs2_asc_desc
}

